add unit test that proves the issue with allowed flow incorrect edges
This commit also makes the assertion a bit nicer. It was difficult to figure
out why the test was failing originally because `.ElementsMatch` print the
complete slices when the assertion fails. For slices with nested object, like
Network Flows, it becomes very difficult to compare side-by-side which
properties didn't match. By looping on each value and checking for equallity,
we're able to get a nice diff on the non-matched object. This only works here
because the graph returned relies on the indexs of the slice to find the edges.